fix: reduce duplicates in `CreateProfile` and `EditProfile`
