+++ app/src/main/java/com/android/periodpals/ui/components/ProfileComponents.kt
@@ -1,5 +1,9 @@
 package com.android.periodpals.ui.components
 
+import android.content.Context
+import android.icu.util.GregorianCalendar
+import android.util.Log
+import android.widget.Toast
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -12,6 +16,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -22,7 +27,11 @@ import androidx.compose.ui.focus.onFocusEvent
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.text.style.TextAlign
+import com.android.periodpals.model.user.User
+import com.android.periodpals.model.user.UserViewModel
 import com.android.periodpals.resources.C.Tag.ProfileScreens
+import com.android.periodpals.ui.navigation.NavigationActions
+import com.android.periodpals.ui.navigation.Screen
 import com.android.periodpals.ui.theme.dimens
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
@@ -56,10 +65,11 @@ fun ProfilePicture(model: Any?, onClick: (() -> Unit)? = null) {
       contentDescription = "profile picture",
       contentScale = ContentScale.Crop,
       modifier =
-          Modifier.size(MaterialTheme.dimens.profilePictureSize)
-              .clip(shape = CircleShape)
-              .then(if (onClick != null) Modifier.clickable { onClick() } else Modifier)
-              .testTag(ProfileScreens.PROFILE_PICTURE),
+      Modifier
+          .size(MaterialTheme.dimens.profilePictureSize)
+          .clip(shape = CircleShape)
+          .then(if (onClick != null) Modifier.clickable { onClick() } else Modifier)
+          .testTag(ProfileScreens.PROFILE_PICTURE),
   )
 }
 
@@ -73,10 +83,11 @@ fun ProfilePicture(model: Any?, onClick: (() -> Unit)? = null) {
 fun ProfileSection(text: String, testTag: String) {
   Text(
       modifier =
-          Modifier.fillMaxWidth()
-              .padding(top = MaterialTheme.dimens.small2)
-              .wrapContentHeight()
-              .testTag(testTag),
+      Modifier
+          .fillMaxWidth()
+          .padding(top = MaterialTheme.dimens.small2)
+          .wrapContentHeight()
+          .testTag(testTag),
       text = text,
       textAlign = TextAlign.Start,
       style = MaterialTheme.typography.titleSmall,
@@ -94,10 +105,11 @@ fun ProfileInputName(name: String, onValueChange: (String) -> Unit) {
   var isFocused by remember { mutableStateOf(false) }
   OutlinedTextField(
       modifier =
-          Modifier.fillMaxWidth()
-              .wrapContentHeight()
-              .testTag(ProfileScreens.NAME_INPUT_FIELD)
-              .onFocusEvent { focusState -> isFocused = focusState.isFocused },
+      Modifier
+          .fillMaxWidth()
+          .wrapContentHeight()
+          .testTag(ProfileScreens.NAME_INPUT_FIELD)
+          .onFocusEvent { focusState -> isFocused = focusState.isFocused },
       value = name,
       onValueChange = onValueChange,
       textStyle = MaterialTheme.typography.labelLarge,
@@ -124,10 +136,11 @@ fun ProfileInputDob(dob: String, onValueChange: (String) -> Unit) {
   var isFocused by remember { mutableStateOf(false) }
   OutlinedTextField(
       modifier =
-          Modifier.fillMaxWidth()
-              .wrapContentHeight()
-              .testTag(ProfileScreens.DOB_INPUT_FIELD)
-              .onFocusEvent { focusState -> isFocused = focusState.isFocused },
+      Modifier
+          .fillMaxWidth()
+          .wrapContentHeight()
+          .testTag(ProfileScreens.DOB_INPUT_FIELD)
+          .onFocusEvent { focusState -> isFocused = focusState.isFocused },
       value = dob,
       onValueChange = onValueChange,
       textStyle = MaterialTheme.typography.labelLarge,
@@ -154,10 +167,11 @@ fun ProfileInputDescription(description: String, onValueChange: (String) -> Unit
   var isFocused by remember { mutableStateOf(false) }
   OutlinedTextField(
       modifier =
-          Modifier.fillMaxWidth()
-              .wrapContentHeight()
-              .testTag(ProfileScreens.DESCRIPTION_INPUT_FIELD)
-              .onFocusEvent { focusState -> isFocused = focusState.isFocused },
+      Modifier
+          .fillMaxWidth()
+          .wrapContentHeight()
+          .testTag(ProfileScreens.DESCRIPTION_INPUT_FIELD)
+          .onFocusEvent { focusState -> isFocused = focusState.isFocused },
       value = description,
       onValueChange = onValueChange,
       textStyle = MaterialTheme.typography.labelLarge,
@@ -177,18 +191,99 @@ fun ProfileInputDescription(description: String, onValueChange: (String) -> Unit
 }
 
 /**
- * A composable function that displays a save button with [onClick] as the action to be executed
- * when the button is clicked.
+ * A composable function that displays a save button and attempts to save the user data
  *
- * @param onClick The action to be executed when the button is clicked.
+ * @param name The name entered by the user.
+ * @param dob The date of birth entered by the user.
+ * @param description The description entered by the user.
+ * @param context The context used to show Toast messages.
+ * @param profileImageUri The URI of the profile image selected by the user.
+ * @param userViewModel The ViewModel that handles user data.
+ * @param userState The current state of the user.
+ * @param navigationActions The navigation actions to navigate between screens.
  */
 @Composable
-fun ProfileSaveButton(onClick: () -> Unit) {
-  Button(
-      modifier = Modifier.wrapContentSize().testTag(ProfileScreens.SAVE_BUTTON),
-      onClick = onClick,
+fun ProfileSaveButton(
+    name: String,
+    dob: String,
+    description: String,
+    profileImageUri: String,
+    context: Context,
+    userViewModel: UserViewModel,
+    userState: State<User?>,
+    navigationActions: NavigationActions
+) {
+
+    Button(
+      modifier = Modifier
+          .wrapContentSize()
+          .testTag(ProfileScreens.SAVE_BUTTON),
+        onClick = {
+            val errorMessage = validateFields(name, dob, description)
+            if (errorMessage != null) {
+                Log.d(LOG_TAG, "$LOG_FAILURE: $errorMessage")
+                Toast.makeText(context, errorMessage, Toast.LENGTH_SHORT).show()
+            }
+
+            Log.d(LOG_TAG, LOG_SAVING_PROFILE)
+            val newUser =
+                User(
+                    name = name,
+                    dob = dob,
+                    description = description,
+                    imageUrl = profileImageUri
+                )
+            userViewModel.saveUser(newUser)
+            if (userState.value == null) {
+                Log.d(LOG_TAG, LOG_FAILURE)
+                Toast.makeText(context, TOAST_FAILURE, Toast.LENGTH_SHORT).show()
+            }
+
+            Log.d(LOG_TAG, LOG_SUCCESS)
+            Toast.makeText(context, TOAST_SUCCESS, Toast.LENGTH_SHORT).show()
+            navigationActions.navigateTo(Screen.PROFILE)
+        },
       enabled = true,
   ) {
     Text(text = SAVE_BUTTON_TEXT, style = MaterialTheme.typography.bodyMedium)
   }
+}
+
+/**
+ * Validates the fields of the screen.
+ *
+ * @param name The name entered by the user.
+ * @param dob The date of birth entered by the user.
+ * @param description The description entered by the user.
+ * @return An error message if validation fails, otherwise null.
+ */
+fun validateFields(name: String, dob: String, description: String): String? {
+    return when {
+        !validateDate(dob) -> ERROR_INVALID_DATE
+        name.isEmpty() -> ERROR_INVALID_NAME
+        description.isEmpty() -> ERROR_INVALID_DESCRIPTION
+        else -> null
+    }
+}
+
+/**
+ * Validates the date is in the format DD/MM/YYYY and is a valid date.
+ *
+ * @param date The date string to validate.
+ * @return True if the date is valid, otherwise false.
+ */
+fun validateDate(date: String): Boolean {
+    val parts = date.split("/")
+    val calendar = GregorianCalendar.getInstance()
+    calendar.isLenient = false
+    if (parts.size == 3) {
+        return try {
+            calendar.set(parts[2].toInt(), parts[1].toInt() - 1, parts[0].toInt())
+            calendar.time
+            true
+        } catch (e: Exception) {
+            false
+        }
+    }
+    return false
 }
